% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleotide_diversity_species.R
\name{align_sequence}
\alias{align_sequence}
\title{function to align sequences with DECIPHER}
\usage{
align_sequence(sequences, MinimumNumberOfSequencesBySpecies)
}
\description{
function to align sequences with DECIPHER
}
