% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_bold_res.R
\name{cluster_classification_bold_res}
\alias{cluster_classification_bold_res}
\title{Cluster classification}
\usage{
cluster_classification_bold_res(bold_res)
}
\arguments{
\item{bold_res}{a data.frame with same fields than resBold$data$markercode and a supplementary column with DNA sequences as string}
}
\value{
a data.frame of number of species and number of individual by species
}
\description{
Number of potentially informative intraspecific clusters
}
