% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grid.R
\name{color_GD}
\alias{color_GD}
\title{Colors of numeric value}
\usage{
color_GD(gd, limites = c(0.001, 0.002, 0.003, 0.005, 0.01, 0.025))
}
\arguments{
\item{gd}{list of numeric values}

\item{limites}{numeric vector of numeric limits values
default value = c(0.001,0.002,0.003,0.005,0.01,0.025)}
}
\value{
a list of color hex value corresponding to the interval of each numeric value
}
\description{
attributes a hex color value to a numeric value based on
intervals of numeric value
}
