% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_bold_res.R
\name{date_bold_res}
\alias{date_bold_res}
\title{run date of bold specimen}
\usage{
date_bold_res(bold_res)
}
\arguments{
\item{bold_res}{a data.frame with same fields than resBold$data$markercode and a supplementary column with DNA sequences as string.}
}
\value{
a data.frame with years and cumulative number of BOLD specimenby year
}
\description{
from the prepared data.fram of query BOLD dataset, the function returns a data.frame of number of cumulative contribtion
of new specimen for each year
}
