% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_bold_res.R
\name{extract_date_from_str}
\alias{extract_date_from_str}
\title{extract date from a string}
\usage{
extract_date_from_str(datestr)
}
\arguments{
\item{datestr}{a string of a list of dates}
}
\value{
a string of a year
}
\description{
From a string formated as follow: "2006-01-26 19:24:04|2006-01-27 01:29:30". This function return the first year.
}
