% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishbase_verify_bold_species.R
\name{fishbase_names_check}
\alias{fishbase_names_check}
\title{Check for alternate versions of a scientific name and return the scientific names FishBase recognizes as valid}
\usage{
fishbase_names_check(species_name)
}
\arguments{
\item{a}{string of scientific names}
}
\value{
a string of the validated names or NA if not found in FishBase
}
\description{
This is the `validate_names` function from FishBase package
}
