% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishbase_verify_bold_species.R
\name{fishbase_verify_bold_species}
\alias{fishbase_verify_bold_species}
\title{Verify against FishBase the species names of BOLD specimen}
\usage{
fishbase_verify_bold_species(datGeoSpecMarkSeqClean)
}
\arguments{
\item{datGeoSpecMarkSeqClean}{A curated dataframe generated by `prepare_bold_res` function from rgeogendiv package
 with 80 columns including:
`species_name`: this column is a list of string object. The only one necessary to validate species name on FishBase.}
}
\value{
A data.frame with a supplementary column `fishbase_species_name`.
This column is a list of strings.
It gives synonyms of species names used by FishBase to describe BOLD specimen.
If the specimen BOLD species name is not found in FishBase, then a NA value is given.
}
\description{
Add a column named `fishbase_species_name` to a curated BOLD specimen dataframe.
This column gives the species name used to identify the specimen in FishBase
If the specimen BOLD species name is not found in FishBase, then a NA value is given.
}
