% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_spatialpolygons.R
\name{grid_spatialpolygons}
\alias{grid_spatialpolygons}
\title{Build a map grid spatialpolygons}
\usage{
grid_spatialpolygons(
  siteSize = 260000,
 
    projectionCRS = "+proj=cea +lon_0=0 +lat_ts=30 +x_0=0 +y_0=0 +datum=WGS84 +ellps=WGS84 +units=m +no_defs"
)
}
\arguments{
\item{siteSize}{size of each cell of the grid in meter}

\item{projectionCRS}{Interface class to the PROJ projection and transformation system.
Default value is "+proj=cea +lon_0=0 +lat_ts=30 +x_0=0 +y_0=0 +datum=WGS84 +ellps=WGS84 +units=m +no_defs"}
}
\value{
a spatialpolygons object of all sites of the grid into the projection `projectionCRS` given as argument
}
\description{
The function download from NaturalEarth database a worldmap raster.
 A grid is built from this raster.
 Then it returns a spatialpolygons object from `sp` package of each site of the grid.
}
