% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grid.R
\name{legend_limit_GD}
\alias{legend_limit_GD}
\title{labels of Limits}
\usage{
legend_limit_GD(limites = c(0.001, 0.002, 0.003, 0.005, 0.01, 0.025))
}
\arguments{
\item{limites}{numeric vector of numeric limits
default value = c(0.001,0.002,0.003,0.005,0.01,0.025)}
}
\value{
a list of intervals labels
}
\description{
A list of label of intervals is produced based on a
list of limits.
For instance if you provide list of limits [0.1, 0.2, 0.3, 1]
It will returns a list ['< 0.1', '0.1 - 0.2', '0.2 - 0.3', '> 1'].
}
