% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grid.R
\name{legend_qt_GD}
\alias{legend_qt_GD}
\title{labels of Intervals}
\usage{
legend_qt_GD(gd, intervalles = c(0, 0.1, 0.2, 0.3, 0.4, 0.6, 1))
}
\arguments{
\item{gd}{list of genetic diversity values}

\item{intervalles}{numeric vector of probabilities with values in [0,1].}
}
\value{
a list of interval labels
}
\description{
A list of label of intervals is produced based on a
numeric vector of probabilities with values in [0,1].
The smallest observation corresponds to a probability of 0
 and the largest to a probability of 1.
}
