% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_bold_res.R
\name{marker_bold_res}
\alias{marker_bold_res}
\title{Return a data.frame of markers and count from BOLD dataset}
\usage{
marker_bold_res(bold_res)
}
\arguments{
\item{bold_res}{a data.frame with same fields than resBold$data$markercode and a supplementary column with DNA sequences as string.}
}
\value{
a table of makercode and number of occurences into the BOLD dataset
}
\description{
after preparing bold dataset query using `prepare_bold_res` function, this function will extract a data.frame of markers.
}
