% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleotide_diversity_grid.R
\name{nucleotide_diversity_grid}
\alias{nucleotide_diversity_grid}
\title{nucleotide diversity by site on the worldmap grid}
\usage{
nucleotide_diversity_grid(nucdivSites, grid.sp)
}
\arguments{
\item{nucdivSites}{a data.frame of nucleotide diversity by site and site grid's ids}

\item{grid.sp}{a spatialpolygons object of all sites of the grid}
}
\value{
a data.frame of nucleotide diversity by site
}
\description{
The function assign to each polygons of a grid the value of nucleotide diversity
as it is indicated in the data.frame `nucdivSites`.
}
