% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleotide_diversity_sites.R
\name{nucleotide_diversity_sites}
\alias{nucleotide_diversity_sites}
\title{nucleotide diversity by site}
\usage{
nucleotide_diversity_sites(nucdivSpecies)
}
\arguments{
\item{nucdivSpecies}{a data.frame of nucleotide diversity by species
generated by `species_nucleotide_diversity` function.}
}
\value{
a data.frame of nucleotide diversity by site
}
\description{
nucleotide diversity by site
}
