% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleotide_diversity_species.R
\name{nucleotide_diversity_species}
\alias{nucleotide_diversity_species}
\title{nucleotide diversity by species applied to all sites}
\usage{
nucleotide_diversity_species(
  specimen.df,
  sequenceIntersectSites,
  MinimumNumberOfSequencesBySpecies,
  fishbaseValidation = FALSE
)
}
\arguments{
\item{specimen.df}{a data.frame of specimen}

\item{sequenceIntersectSites}{a matrix of presence/absence TRUE/FALSE of a specimen within a site
with column as site ID
and row as specimen spatial points ID}

\item{MinimumNumberOfSequencesBySpecies}{an integer minimum number of sequences representatives of a species in a site}
}
\value{
a dataframe with taxonomy and nucleotide diversity and number of sequences by species
species with not enough sequences to calculate nucleotide diversity are removed
}
\description{
nucleotide diversity by species applied to all sites
}
