% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleotide_diversity_species.R
\name{number_sequences_by_points}
\alias{number_sequences_by_points}
\title{Number of sequences in a site}
\usage{
number_sequences_by_points(sequenceIntersectSites)
}
\arguments{
\item{sequenceIntersectSites}{a matrix of presence/absence TRUE/FALSE of a specimen within a site
with column as site ID}
}
\value{
a list with the number of sequences within a site
}
\description{
Number of sequences in a site
}
