% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grid.R
\name{plot_grid}
\alias{plot_grid}
\title{Print worldmap of mean species genetic diversity}
\usage{
plot_grid(nucdivGrid, limites = c(0.001, 0.002, 0.003, 0.005, 0.01, 0.025))
}
\arguments{
\item{nucdivGrid}{a data.frame of nucleotide diversity by site}

\item{limites}{numeric vector of numeric limits values
default value = c(0.001,0.002,0.003,0.005,0.01,0.025)}
}
\value{
a `ggplot2` object of the worldmap of mean species genetic diversity
}
\description{
downloads coastlines, rivers and lakes shapefile from `naturalearth`.
print sites of the grid worldmap with colors and labels of their genetic values.
}
