% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_bold_res.R
\name{prepare_bold_res}
\alias{prepare_bold_res}
\title{Prepare raw BOLD specimen and DNA sequences dataset}
\usage{
prepare_bold_res(
  bold_res,
  marker_code = "",
  species_names = TRUE,
  coordinates = TRUE,
  ambiguities = TRUE,
  min_length = 0,
  max_length = 800
)
}
\arguments{
\item{bold_res}{A list of lists returned by bold_seqspec command from bold package. They are 2 lists:
* Specimen information (spatial coordinates, taxonomy...)
* DNA barcode sequences}

\item{marker_code}{If not empty, only specimen with field data$markercode matching marker_code names are kept. By default, markercode filtering is not applied.}

\item{species_names}{a logical value indicating whether specimen with no species name information should be removed.}

\item{coordinates}{a logical value indicating whether specimen with no latitude or longitude spatial coordinates information should be removed.}

\item{ambiguities}{a logical value indicating whether specimen with DNA sequence containing IUPAC ambiguities should be removed.}

\item{min_length}{numeric. Minimum length of a DNA sequence. Default value: 0 bp.}

\item{max_length}{numeric. Maximum length of a DNA sequence. Default value: 800 bp}
}
\value{
a data.frame with same fields than resBold$data$markercode and a supplementary column with DNA sequences as string.
}
\description{
Filter and mutate BOLD dataset to produce a curated data.frame
 with rows as individual specimen and columns as specimen information.
 It adds a new column sequence with fasta sequences as string.
}
