% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialpoints_bold_df.R
\name{spatialpoints_bold_df}
\alias{spatialpoints_bold_df}
\title{Transform curated BOLD data.frame into spatialpoints object from `sp` package}
\usage{
spatialpoints_bold_df(
  dfBold,
  projectionCRS = "+init=epsg:3347",
  fishbaseValidation = FALSE
)
}
\arguments{
\item{dfBold}{A data.frame with specimen sequences as row and 81 descriptors as column.
Mandatory fields:
lon,
lat,
species_name,
fishbase_species_name,
genus_name,
family_name,
order_name,
class_name,
sequence}

\item{projectionCRS}{Interface class to the PROJ projection and transformation system.
Default value is "+init=epsg:3347"}
}
\value{
a spatialpoints object with taxonomic information and DNA sequence
}
\description{
Given a BOLD individual georeferenced sequences dataframe and a projection CRS,
this function will return a spatialpoints object with all coordinates of BOLD specimen.
}
