% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleotide_diversity_species.R
\name{species_nucleotide_diversity_site}
\alias{species_nucleotide_diversity_site}
\title{Nucleotide diversity by species}
\usage{
species_nucleotide_diversity_site(
  site.ids,
  specimen.df,
  sequenceIntersectSites,
  MinimumNumberOfSequencesBySpecies
)
}
\arguments{
\item{site.ids}{index of the site of the grid}

\item{specimen.df}{a data.frame with taxonomy+sequence dna information}

\item{sequenceIntersectSites}{a matrix of presence/absence TRUE/FALSE of a specimen within a site
with column as site ID
and row as specimen spatial points ID}

\item{MinimumNumberOfSequencesBySpecies}{an integer minimum number of sequences representatives of a species in a site}
}
\value{
a data.frame with number of individuals and nucleotide diversity for each species in the site.
}
\description{
nucleotide diversity by species
arguments:
* specimen.df a dataframe with taxonomy+sequence dna information
* buffer.Within a dataframe with row as indiv seq and col as rls pts
with TRUE/FALSE presence/absence id of sequence specimen.df within a given buffer around a given rls pts
* rls.pts.index index of the RLS pts
it calculates
* alignment of sequences by species
* number of sequences by species
* nucleotide diversity
it returns
a dataframe with taxonomy and nucleotide diversity and number of sequences by species
species with not enough sequences to calculate nucleotide diversity are removed
}
