% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen_intersect_site.R
\name{specimen_intersect_site}
\alias{specimen_intersect_site}
\title{specimen spatial points within spatial polygons site of grid}
\usage{
specimen_intersect_site(
  specimen.df,
  grid.sp,
  projectionCRS = sp::CRS(sp::proj4string(grid.sp))@projargs[1]
)
}
\arguments{
\item{specimen.df}{a data.frame of specimen}

\item{grid.sp}{a spatialpolygons object of all sites of the grid}

\item{projectionCRS}{Interface class to the PROJ projection and transformation system.
Default value is the projection of the `grid.sp` parameter.}
}
\value{
a matrix of presence/absence TRUE/FALSE of a specimen within a site
with column as site ID
and row as specimen spatial points ID
}
\description{
specimen spatial points within spatial polygons site of grid
}
